Outsourced decompression methods from CriWare to extra classes.
